tests/wpt/tests/html/semantics/embedded-content/the-iframe-element/iframe_top_navigation_without_user_gesture_nested_same_site.tentative.html

<!doctype html>

Check that a cross-site iframe can't navigate its top window if the user hasn't
interacted with it, even if its parent window is same-site.

<button>Gain user activation</button>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
  async_test(function (t) {
    // Gain user activation to be able to open a popup. This is needed, as we have
    // enabled popup blocking for this test.
    test_driver.click(document.querySelector("button")).then(() => {
      let beforeunloadFired = false;

      window.addEventListener(
        "message",
        t.step_func_done(function (e) {
          assert_equals(
            e.data,
            "ERROR",
            "The child should receive an error when trying to navigate"
          );
          assert_equals(
            beforeunloadFired,
            false,
            "No beforeunload event should have been fired in the parent"
          );
        })
      );

      let parent = open(
        "support/iframe-that-performs-top-navigation-grandparent.sub.html"
      );
      parent.addEventListener("beforeunload", () => {
        beforeunloadFired = true;
      });
    });
  });
</script>
